feat: refactor homepages
